+++ Imaging/vtkImageButterworthHighPass.h
@@ -18,7 +18,7 @@
 // frequency domain by a vtkImageFFT filter.  A vtkImageRFFT filter
 // can be used to convert the output back into the spatial domain.
 // vtkImageButterworthHighPass  the frequency components around 0 are
-// attenuated.  Input and output are in floats, with two components
+// attenuated.  Input and output are in doubles, with two components
 // (complex numbers).
 // out(i, j) = 1 / (1 + pow(CutOff/Freq(i,j), 2*Order));
 
@@ -42,15 +42,15 @@ public:
   // Set/Get the cutoff frequency for each axis.
   // The values are specified in the order X, Y, Z, Time.
   // Units: Cycles per world unit (as defined by the data spacing).
-  vtkSetVector3Macro(CutOff,float);
-  void SetCutOff(float v) {this->SetCutOff(v, v, v);}
-  void SetXCutOff(float v);
-  void SetYCutOff(float v);
-  void SetZCutOff(float v);
-  vtkGetVector3Macro(CutOff,float);
-  float GetXCutOff() {return this->CutOff[0];}
-  float GetYCutOff() {return this->CutOff[1];}
-  float GetZCutOff() {return this->CutOff[2];}
+  vtkSetVector3Macro(CutOff,double);
+  void SetCutOff(double v) {this->SetCutOff(v, v, v);}
+  void SetXCutOff(double v);
+  void SetYCutOff(double v);
+  void SetZCutOff(double v);
+  vtkGetVector3Macro(CutOff,double);
+  double GetXCutOff() {return this->CutOff[0];}
+  double GetYCutOff() {return this->CutOff[1];}
+  double GetZCutOff() {return this->CutOff[2];}
 
   // Description:
   // The order determines sharpness of the cutoff curve.
@@ -62,7 +62,7 @@ protected:
   ~vtkImageButterworthHighPass() {};
 
   int Order;
-  float CutOff[3];
+  double CutOff[3];
   
   void ThreadedExecute(vtkImageData *inData, vtkImageData *outData,
                        int outExt[6], int id);
